Add hip warpsize retrieval in common header
